Moved to Java version 1.6 and added maven profile for linux building

+++ mockrunner-core/src/main/java/com/mockrunner/util/common/ArrayUtil.java
@@ -452,7 +452,7 @@ public class ArrayUtil
             if (null == currentValue) {
                 nameMap.put(name, 1);
             } else {
-                nameMap.put(name, currentValue.intValue() + 1);
+                nameMap.put(name, currentValue + 1);
             }
         }
         return nameMap;
